fix(ci): AtomicU64→AtomicUsize for MIPS32, disable fail-fast
- ubus_client.rs: use AtomicUsize instead of AtomicU64 (unavailable on
  MIPS 32-bit targets which lack 64-bit atomics)
- build-package.yml: fail-fast: false so one arch failure doesn't
  cancel x86_64/arm64 builds

+++ crates/tollgate-net/src/v1/server/ubus_client.rs
@@ -37,7 +37,7 @@
 //! `result[0]` is the status code (0 = success). `result[1]` is the
 //! response payload.
 
-use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::atomic::{AtomicUsize, Ordering};
 use std::time::Duration;
 
 use serde::{Deserialize, Serialize};
@@ -119,7 +119,7 @@ pub struct UbusClient {
     /// HTTP request timeout.
     timeout: Duration,
     /// JSON-RPC request ID counter.
-    id_counter: AtomicU64,
+    id_counter: AtomicUsize,
     /// HTTP client.
     client: reqwest::Client,
 }
@@ -135,7 +135,7 @@ impl UbusClient {
             url,
             token: None,
             timeout: DEFAULT_TIMEOUT,
-            id_counter: AtomicU64::new(0),
+            id_counter: AtomicUsize::new(0),
             client: reqwest::Client::builder()
                 .timeout(DEFAULT_TIMEOUT)
                 .build()
@@ -378,7 +378,7 @@ impl UbusClient {
         method: &str,
         params: serde_json::Value,
     ) -> Result<serde_json::Value, UbusError> {
-        let id = self.id_counter.fetch_add(1, Ordering::Relaxed) + 1;
+        let id = self.id_counter.fetch_add(1, Ordering::Relaxed) as u64 + 1;
         let request = JsonRpcRequest {
             jsonrpc: "2.0",
             id,
